requirements.txt
pydantic~=2.10.6
requests~=2.32.3
pydantic-settings~=2.7.1
aiohttp~=3.11.11
pyTelegramBotAPI~=4.26.0
aiosqlite~=0.20.0
typing_extensions~=4.12.2
tortoise-orm~=0.24.0